Modifications on Sage Macros

+++ Sage/Noulith Recall.js
@@ -4,26 +4,36 @@ const characterToken = ownedTokens.find(ownedToken => (
 ));
 
 const targets = game.user.targets;
-targets.filter((target) => (target.name === "Noulith")).forEach(async (target) => {
-  target.document.delete();
+if (targets.first().name === "Kerachole") {
+  const keracholeTarget = targets.first();
+  const recallItem = keracholeTarget.actor.items.find(item => item.name === "Recall");
+  recallItem.use();
+} else {
+  targets.filter((target) => (target.name === "Noulith")).forEach(async (target) => {
+    const noulithHp = target.actor.system.attributes.hp.value;
 
-  const [spawnedCreature] = await warpgate.spawnAt(characterToken.center, "Noulith", { token: { alpha: 0, elevation: 1 } });
+    const [spawnedCreature] = await warpgate.spawnAt(characterToken.center, "Noulith", { token: { alpha: 0, elevation: 1 } });
+    const spawning = new Sequence()
+      .wait(200)
+      .effect()
+      .file("jb2a.swirling_sparkles.01.blue")
+      .atLocation(spawnedCreature)
+      .wait(200)
+      .animation()
+      .on(spawnedCreature)
+      .opacity(1.0)
 
-  const spawning = new Sequence()
-    .wait(200)
-    .effect()
-    .file("jb2a.swirling_sparkles.01.blue")
-    .atLocation(spawnedCreature)
-    .wait(200)
-    .animation()
-    .on(spawnedCreature)
-    .opacity(1.0)
+    await spawning.play();
 
-  await spawning.play();
+    let token = canvas.tokens.placeables.find(t => t.id === spawnedCreature);
+    token.actor.update({ "system.attributes.hp.value": noulithHp })
 
-  // Check for Zoe Nouliths
-  const zoeNoulithTokens = game.canvas.tokens.ownedTokens.filter((token) => (token.name === 'Noulith') && !!token.actor.getFlag('world', 'Zoe'));
-  if (zoeNoulithTokens < 1) {
-    game.dfreds.effectInterface.removeEffect({ effectName: "Zoe", uuid: game.user.character.uuid });
-  }
-})
+    // Check for Zoe Nouliths
+    const zoeNoulithTokens = game.canvas.tokens.ownedTokens.filter((token) => (token.name === 'Noulith') && !!token.actor.getFlag('world', 'Zoe'));
+    if (zoeNoulithTokens < 1) {
+      game.dfreds.effectInterface.removeEffect({ effectName: "Zoe", uuid: game.user.character.uuid });
+    }
+
+    target.document.delete();
+  })
+}
